feat(fallback): add some packet for play state.

+++ protocol/src/main/kotlin/net/miaomoe/blessing/protocol/util/PositionUtil.kt
@@ -21,14 +21,14 @@ import kotlin.math.pow
 
 object PositionUtil {
 
-    fun getLegacyBlockPosition(position: Position) = (position.x shl 12 or position.z shl 8 or position.y)
+    fun getLegacyBlockPosition(position: Position) = (position.x.toInt() shl 12 or position.z.toInt() shl 8 or position.y.toInt())
 
-    fun getModernBlockPosition(position: Position) = (position.x shl 8 or position.z shl 4 or position.y.let { it - (it shr 4 shl 4) })
+    fun getModernBlockPosition(position: Position) = (position.x.toInt() shl 8 or position.z.toInt() shl 4 or position.y.toInt().let { it - (it shr 4 shl 4) })
 
     fun getFallGravity(tick: Int) = -((0.98.pow(tick.toDouble()) - 1) * 3.92)
 
-    fun getLegacySpawnPosition(position: Position) = (position.x and 0x3FFFFFF shl 38) or (position.y and 0xFFF shl 26) or (position.z and 0x3FFFFFF)
+    fun getLegacySpawnPosition(position: Position) = (position.x.toInt() and 0x3FFFFFF shl 38) or (position.y.toInt() and 0xFFF shl 26) or (position.z.toInt() and 0x3FFFFFF)
 
-    fun getModernSpawnPosition(position: Position) = (position.x and 0x3FFFFFF shl 38) or (position.y and 0x3FFFFFF shl 12) or (position.z and 0xFFF)
+    fun getModernSpawnPosition(position: Position) = (position.x.toInt() and 0x3FFFFFF shl 38) or (position.y.toInt() and 0x3FFFFFF shl 12) or (position.z.toInt() and 0xFFF)
 
 }
